Restore favicon to 'latest' version of player-ui.github.io
It seems there was an error in building the webpage source for the 'latest' that didn't happen for 'next'.

In the case of 'latest' (and therefore how it is seen by default in the world), the favicon call is to `favicon.svg` of the 'type="image/x-icon"', which does not render in the browser (at least, not on Chrome on Linux.)

This PR simply reverts (or corrects) that change. I don't know that this PR can be merged to make the change, though, as it's evident these pages are built using an Intuit internal build system (a not uncommon situation -- maintaining an external build service is a PITA, ask me how I know.)

If there is something I can do from inside of Intuit now that I'm inside, let me know and also email karsten_wade@intuit.com to point me in the right direction. Happy to do self-service, it's also a chance for me to have a "new contributor to Player"-experience. :grin: